[Refactor] Move ArrayType to types
