+++ be/src/storage/types.cpp
@@ -40,7 +40,6 @@
 #include "runtime/large_int_value.h"
 #include "runtime/mem_pool.h"
 #include "runtime/time_types.h"
-#include "storage/array_type_info.h"
 #include "storage/collection.h"
 #include "storage/convert_helper.h"
 #include "storage/decimal12.h"
@@ -49,6 +48,7 @@
 #include "storage/olap_type_infra.h"
 #include "storage/tablet_schema.h" // for TabletColumn
 #include "storage/type_traits.h"
+#include "types/array_type_info.h"
 #include "types/date_value.hpp"
 #include "types/map_type_info.h"
 #include "types/struct_type_info.h"
