+++ be/src/types/CMakeLists.txt
@@ -15,6 +15,7 @@
 set(LIBRARY_OUTPUT_PATH "${BUILD_DIR}/src/types")
 
 add_library(Types STATIC
+    array_type_info.cpp
     bitmap_value.cpp
     date_value.cpp
     hll.cpp
