+++ be/src/types/array_type_info.h
